backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
spacy
google-genai
